backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
orjson==3.10.15

# Neo4j and graph
neo4j==5.27.0
networkx==3.4.2

# Embeddings and Vector Search
sentence-transformers==3.3.1
faiss-cpu==1.12.0

# LLM
openai==1.57.4

# Utilities
numpy==1.26.4
scikit-learn==1.6.1

